fix(admin): initialize translations on client
