+++ packages/earth-map/src/App.tsx
@@ -20,11 +20,13 @@
 import React, { useContext } from 'react';
 import { Provider } from 'react-redux';
 
-import { Spinner } from '@marapp/earth-shared';
+import { Spinner, TranslationService } from '@marapp/earth-shared';
 
 import { Auth0Context, useAuth0 } from './auth/auth0';
 import Main from './pages/main';
 import initStore from './store';
+
+TranslationService.init();
 
 const App = () => {
   const { isLoading } = useAuth0();
