+++ packages/earth-shared/src/index.ts
@@ -16,8 +16,6 @@
   CONDITIONS OF ANY KIND, either express or implied. See the License for the
   specific language governing permissions and limitations under the License.
 */
-import './translations/index';
-
 export * from './components';
 export * from './authorization';
 export * from './services';
@@ -27,3 +25,4 @@ export * from './hooks';
 export * from './assets';
 export * from './theme';
 export * from './locales';
+export * from './translations';
