+++ packages/earth-shared/src/translations/index.ts
@@ -1,60 +1,21 @@
 /*
- * Copyright 2018-2020 National Geographic Society
- *
- * Use of this software does not constitute endorsement by National Geographic
- * Society (NGS). The NGS name and NGS logo may not be used for any purpose without
- * written permission from NGS.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
- * this file except in compliance with the License. You may obtain a copy of the
- * License at
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
+  Copyright 2018-2020 National Geographic Society
 
-import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
+  Use of this software does not constitute endorsement by National Geographic
+  Society (NGS). The NGS name and NGS logo may not be used for any purpose without
+  written permission from NGS.
 
-import { Elang } from '../locales';
+  Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+  this file except in compliance with the License. You may obtain a copy of the
+  License at
 
-import translationEN from './en/translation.json';
-import translationES from './es/translation.json';
-import translationFR from './fr/translation.json';
+      https://www.apache.org/licenses/LICENSE-2.0
 
-import translationENAdmin from './en/admin.translation.json';
-import translationESAdmin from './es/admin.translation.json';
-import translationFRAdmin from './fr/admin.translation.json';
+  Unless required by applicable law or agreed to in writing, software distributed
+  under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+  specific language governing permissions and limitations under the License.
+*/
 
-import SessionStorage from '../services/SessionStorage';
-
-const lang = SessionStorage.get('lang') || Elang.EN;
-
-i18n.use(initReactI18next).init({
-  resources: {
-    [Elang.EN]: {
-      translation: translationEN,
-      admin: translationENAdmin,
-    },
-    [Elang.ES]: {
-      translation: translationES,
-      admin: translationESAdmin,
-    },
-    [Elang.FR]: {
-      translation: translationFR,
-      admin: translationFRAdmin,
-    },
-  },
-  lng: lang,
-  fallbackLng: Elang.EN,
-  interpolation: {
-    escapeValue: false,
-  },
-  keySeparator: false,
-});
-
-i18n.on('languageChanged', (lng) => SessionStorage.add('lang', lng));
+export { default as TranslationService } from './TranslationService';
+export * from './TranslationService';
